% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_coordinate.R
\name{bed2feature}
\alias{bed2feature}
\title{Convert bed format to feture format}
\usage{
bed2feature(bed)
}
\arguments{
\item{bed}{bed format data}
}
\value{
feature format data
}
\description{
Convert bed format to feture format
}
