% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_coordinate.R
\name{bed2gr}
\alias{bed2gr}
\title{Convert bed format to gr format}
\usage{
bed2gr(bed)
}
\arguments{
\item{bed}{bed format data}
}
\value{
gr format data
}
\description{
Convert bed format to gr format
}
