% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_coordinate.R
\docType{package}
\name{easyapi}
\alias{easyepi}
\alias{easyepi-package}
\alias{easyapi}
\title{Conversion between different genome coordinate formats}
\description{
Description: Conversion between different genome coordinate formats
}
\details{
Details: There are 3 different genome coordinate formats is this document.
1. gr: GRanges in GenomicRanges package. 1-base include.
2. bed: data.frame, and thefirst three columns of the data.frame
are chrom, start, and end. 0-base exclude.
3. feature: string like chr1:10-20. 0-base exclude.#'
}
\author{
\strong{Maintainer}: Yihang Cheng \email{navyc@126.com} (\href{https://orcid.org/0009-0006-9576-2180}{ORCID})

}
