% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_coordinate.R
\name{feature2bed}
\alias{feature2bed}
\title{Convert feature format to bed format}
\usage{
feature2bed(feature)
}
\arguments{
\item{feature}{feature format data}
}
\value{
bed format data
}
\description{
Convert feature format to bed format
}
\examples{
feature2bed
}
