% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_coordinate.R
\name{feature2gr}
\alias{feature2gr}
\title{Convert feature format to gr format}
\usage{
feature2gr(feature)
}
\arguments{
\item{feature}{feature format data}
}
\value{
gr format data
}
\description{
Convert feature format to gr format
}
