% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{figPercentileMethy}
\alias{figPercentileMethy}
\title{draw the percentile plot}
\usage{
figPercentileMethy(
  target,
  background,
  is.hypo = TRUE,
  p = 0.8,
  ylab = "",
  title = ""
)
}
\arguments{
\item{target}{the values of target samples. -1 will be removed.}

\item{background}{the values of background samples. -1 will bed removed.}

\item{is.hypo}{If the merged methylation level of target samples is less than background samples.}

\item{p}{the percentile of low methylation group. If is.hypo is true, the p-th percentile is caclulated}

\item{ylab}{the ylab}

\item{title}{the title}
}
\value{
None.
}
\description{
draw the percentile plot of target and background samples
}
\examples{
figPercentileMethy
}
