% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerhelper.R
\name{getfindMotifsGenomeResults}
\alias{getfindMotifsGenomeResults}
\title{Extract the results from the findMotifsGenome.pl script}
\usage{
getfindMotifsGenomeResults(result.dir, q.value = NULL, foldchange = NULL)
}
\arguments{
\item{result.dir}{the result directory of  findMotifsGenome.pl}

\item{q.value}{the maximum q-value of the results}

\item{foldchange}{the minimum fold change of the results}
}
\description{
Currently only knownResults.txt results are extracted. Filtering of results is supported
}
\examples{
getfindMotifsGenomeResults('/dmc/homer80/LUAD.hypo',0.05, 1.2)

}
