% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_coordinate.R
\name{gr2bed}
\alias{gr2bed}
\title{Convert gr format to bed format}
\usage{
gr2bed(gr)
}
\arguments{
\item{gr}{gr format data}
}
\value{
bed format data
}
\description{
Convert gr format to bed format
}
