% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_coordinate.R
\name{gr2feature}
\alias{gr2feature}
\title{Convert gr format to feture format}
\usage{
gr2feature(gr)
}
\arguments{
\item{gr}{gr format data}
}
\value{
feature format data
}
\description{
Convert gr format to feture format
}
