% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromt_something.R
\name{percent2numeric}
\alias{percent2numeric}
\title{convert a string percentage into numeric}
\usage{
percent2numeric(x)
}
\arguments{
\item{x}{a percentage string end with \%}
}
\value{
the numeric
}
\description{
convert a string percentage into numeric
}
\examples{
percent2numeric('\%0.01')

0.1
}
